Refactor formapp into separate modules
The main app, utils and components.  This is more aligned with the form builder tool.
