+++ src/js/formapp.js
@@ -8,10 +8,18 @@ import 'sass/formapp/bootstrap.min.css';
 
 import 'sass/formapp-core.scss';
 
-import { extend, map, reduce, debounce } from 'underscore';
+import { extend, map, debounce } from 'underscore';
 import Backbone from 'backbone';
 
 import { versions } from './config';
+
+import {
+  getScriptContext,
+  getSubmission,
+  getChangeReducers,
+} from 'js/formapp/utils';
+
+import 'js/formapp/components';
 
 import PreloadRegion from 'js/regions/preload_region';
 
@@ -26,98 +34,21 @@ function scrollTop() {
   window.scrollTo({ top: 0 });
 }
 
-const NestedComponent = Formio.Components.components.nested;
-
-class Snippet extends NestedComponent {
-  constructor(...args) {
-    super(...args);
-    this.noField = true;
-  }
-  static schema(...extendArgs) {
-    return NestedComponent.schema({
-      label: 'Snippet',
-      key: 'snippet',
-      type: 'snippet',
-      components: [],
-      input: false,
-      persistent: false,
-      snippet: null,
-    }, ...extendArgs);
-  }
-}
-
-Formio.use({
-  components: {
-    snippet: Snippet,
-  },
-});
-
 function getDirectory(directoryName, query) {
   return router.request('fetch:directory', { directoryName, query });
 }
 
-// Note: Allows for setting the submission at form instantiation
-// https://github.com/formio/formio.js/pull/4580
-const webformInit = Formio.Displays.displays.webform.prototype.init;
-
-Formio.Displays.displays.webform.prototype.init = function() {
-  if (this.options.data) {
-    const data = extend({}, this.options.data);
-    this._submission = { data };
-    this._data = data;
-  }
-  webformInit.call(this);
-};
-
-// NOTE: Evaluator should throw errors
-// https://github.com/formio/formio.js/issues/4613
-const evaluator = Formio.Evaluator.evaluator;
-Formio.Evaluator.evaluator = function(func, ...params) {
-  try {
-    return evaluator(func, ...params);
-  } catch (e) {
-    /* eslint-disable-next-line no-console */
-    console.error(e);
-  }
-};
-
-function getScriptContext(contextScripts) {
-  return Formio.createForm(document.createElement('div'), {}).then(form => {
-    const context = reduce(contextScripts, (memo, script) => {
-      return extend({}, memo, FormioUtils.evaluate(script, form.evalContext(memo)));
-    }, { getDirectory });
-
-    form.destroy();
-
-    return context;
-  });
-}
-
-function getSubmission(formData, formSubmission, reducers, evalContext) {
-  return Formio.createForm(document.createElement('div'), {}, { evalContext }).then(form => {
-    const submission = reduce(reducers, (memo, reducer) => {
-      return FormioUtils.evaluate(reducer, form.evalContext({ formSubmission: memo, formData })) || memo;
-    }, formSubmission);
-
-    form.destroy();
-
-    return submission;
-  });
+function getContext(contextScripts) {
+  return getScriptContext(contextScripts, { getDirectory });
 }
 
 let prevSubmission;
 let isChanging = false;
 
-const hasChangedFunction = 'return function hasChanged(key) { return !_.isEqual(_.get(formSubmission, key), _.get(prevSubmission, key)); }';
-
 const onChange = debounce(function(form, changeReducers) {
   isChanging = true;
 
-  const data = reduce(changeReducers, (memo, reducer) => {
-    const context = form.evalContext({ formSubmission: memo, prevSubmission });
-    context.hasChanged = FormioUtils.evaluate(hasChangedFunction, context);
-    return FormioUtils.evaluate(reducer, context) || memo;
-  }, structuredClone(form.submission.data));
+  const data = getChangeReducers(form, changeReducers, structuredClone(form.submission.data), prevSubmission);
 
   form.setSubmission({ data });
 
@@ -127,7 +58,7 @@ const onChange = debounce(function(form, changeReducers) {
 }, 300);
 
 async function renderForm({ definition, formData, formSubmission, reducers, changeReducers, contextScripts, beforeSubmit }) {
-  const evalContext = await getScriptContext(contextScripts);
+  const evalContext = await getContext(contextScripts);
 
   const submission = await getSubmission(formData, formSubmission, reducers, evalContext);
   prevSubmission = structuredClone(submission);
@@ -189,7 +120,7 @@ async function renderForm({ definition, formData, formSubmission, reducers, chan
 }
 
 async function renderPreview({ definition, contextScripts }) {
-  const evalContext = await getScriptContext(contextScripts);
+  const evalContext = await getContext(contextScripts);
 
   extend(evalContext, { isPreview: true });
 
@@ -197,7 +128,7 @@ async function renderPreview({ definition, contextScripts }) {
 }
 
 async function renderResponse({ definition, formSubmission, contextScripts }) {
-  const evalContext = await getScriptContext(contextScripts);
+  const evalContext = await getContext(contextScripts);
 
   extend(evalContext, { isResponse: true });
